requirements.txt
django
djangorestframework
psycopg2-binary
drf-yasg
gunicorn
django-environ
django-filter
python-decouple
requests